refactor: move TodoColors class to 'themes' package
